🐛 fix: sessionProvider에 불필요한 hook, session 코드 삭제 #53
